Fixed Charge issue in TGRSIDetectorHit

+++ libraries/TGRSIAnalysis/TGRSIDetector/TGRSIDetectorHit.cxx
@@ -38,7 +38,7 @@ double TGRSIDetectorHit::GetEnergy(Option_t *opt) const{
       printf("no TChannel set for this address\n");
       return 0.00;
    }
-   return chan->CalibrateENG((int)GetCharge());
+   return chan->CalibrateENG(GetCharge());
 }
 
 Double_t TGRSIDetectorHit::SetEnergy(Option_t *opt) {
